+++ client-app/src/admin/tests/cube/CubeTestModel.ts
@@ -1,10 +1,10 @@
 import {ChartModel} from '@xh/hoist/cmp/chart';
-import {GridModel, timeCol, TreeStyle} from '@xh/hoist/cmp/grid';
+import {GridModel, time, TreeStyle} from '@xh/hoist/cmp/grid';
 import {fragment} from '@xh/hoist/cmp/layout';
 import {HoistModel, managed, XH} from '@xh/hoist/core';
 import {numberEditor, textEditor} from '@xh/hoist/desktop/cmp/grid';
 import {fmtNumber, numberRenderer} from '@xh/hoist/format';
-import { action, bindable, comparer, observable, runInAction } from '@xh/hoist/mobx';
+import {action, bindable, comparer, observable, runInAction} from '@xh/hoist/mobx';
 import {wait} from '@xh/hoist/promise';
 import {SECONDS} from '@xh/hoist/utils/datetime';
 import {forEach, isEmpty} from 'lodash';
@@ -386,7 +386,7 @@ export class CubeTestModel extends HoistModel {
                 {
                     field: 'time',
                     editable: false,
-                    ...timeCol
+                    ...time
                 }
             ]
         });